[PM-2966] Move networking to separate top module
